+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## UNRELEASED
+
+IMPROVEMENTS:
+ * plugins/target/aws-asg: Support EC2 role credentials [[GH-564](https://github.com/hashicorp/nomad-autoscaler/pull/564)]
+
 ## 0.3.5 (January 20, 2022)
 
 IMPROVEMENTS: